requirements.txt
# Data & APIs
pandas==2.2.0
alpaca-py==0.22.0
yfinance==0.2.36
finnhub-python==2.4.19

# AI
google-generativeai==0.8.3

# Database
# (sqlite3 is built into Python)

# Scheduling & File Watching
watchdog==4.0.0
pytz==2024.1

# Configuration
pyyaml==6.0.1

# Logging
colorlog==6.8.2

# Testing
pytest==8.0.0
